refactor: extract subscription primitives to aiounifi/models/subscription.py
Move ItemEvent, CallbackType, SubscriptionType, UnsubscribeType,
ID_FILTER_ALL, and SubscriptionHandler to a new shared module so both
the legacy interfaces stack and the Network API v1 stack can import from
a neutral location without a cross-package dependency.

Re-export all symbols from interfaces/api_handlers.py for backwards
compatibility. Update network/v1/api_handlers.py to import from the
shared module instead of the legacy interfaces package.
